refactor(tag): extract `TagBody` into imported component
