+++ packages/components/tag/README.md
@@ -63,7 +63,7 @@ export default Example;
 | `styles.body`          | `object`                                                                     |          |            |                                                                                                                          |
 | `linkTo`               | `string`                                                                     |          |            | Link of the tag when not disabled                                                                                        |
 | `isDisabled`           | `bool`                                                                       |          | `false`    | Disable the tag element along with the option to remove it.                                                              |
-| `onRemove`             | `func`                                                                       |          |            | Called when remove button is clicked. Signature: `(event) => void`                                                       |
+| `onRemove`             | `func`                                                                       |          |            | Called when remove button is clicked.<br /> Signature: `(event) => void`                                                 |
 | `onClick`              | `func`                                                                       |          |            | Called when tag element is clicked. This is not called when remove button is clicked.<br /> Signature: `(event) => void` |
 | `horizontalConstraint` | `enum`<br>Possible values:<br>`'xs' \| 's' \| 'm' \| 'l' \| 'xl' \| 'scale'` |          | `'scale'`  | Horizontal size limit of the input field.                                                                                |
 | `children`             | `node`                                                                       |    ✅    |            | Content rendered within the tag                                                                                          |
